Question reponse answer delete (#17)
* answer entity created and question controller optimized

* answers ok, image upload ok, answer delete ok

+++ src/src/Controller/AdvertController.php
@@ -4,10 +4,13 @@ namespace App\Controller;
 
 use App\Entity\Advert;
 use App\Entity\Question;
+use App\Entity\Answer;
 use App\Form\AdvertType;
 use App\Form\QuestionType;
 use App\Form\SearchAdvertType;
+use App\Form\AnswerType;
 use App\Repository\AdvertRepository;
+use App\Repository\AnswerRepository;
 use App\Repository\QuestionRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -21,6 +24,7 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 
 class AdvertController extends AbstractController
 {
+    
     #[Route('/advert', name: 'app_advert')]
     public function index(AdvertRepository $advertRepository, Request $request)
     {
@@ -47,61 +51,80 @@ class AdvertController extends AbstractController
             $form->handleRequest($request);
             if ($form->isSubmitted() && $form->isValid()) {
                 $image = $form->get('image')->getData();
-                // this condition is needed because the 'brochure' field is not required
-                // so the PDF file must be processed only when a file is uploaded
                 if ($image) {
                     $originalFilename = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
-                    // this is needed to safely include the file name as part of the URL
                     $safeFilename = $slugger->slug($originalFilename);
                     $newFilename = $safeFilename.'-'.uniqid().'.'.$image->guessExtension();
                     
-                    // Move the file to the directory where brochures are stored
                     try {
                         $image->move(
                             $this->getParameter('images_directory'),
                             $newFilename
                         );
                     } catch (FileException $e) {
-                        // ... handle exception if something happens during file upload
+                        dd($e);
                     }
                     $advert->setImageFileName($newFilename);
                 }
-                
-                // ... persist the $product variable or any other work
-                
+                                
                 $advert = $form->getData();
                 $advert->setUserRegister($user);
                 $advertRepository->save($advert, true);
     
                 return $this->redirectToRoute('app_advert');
             }
-            
-            return $this->render('advert/addAdvert.html.twig', [
-                'form' => $form->createView()
-            ]);
         }else{
             return $this->redirectToRoute('app_advert');
         }
         
+        return $this->render('advert/addAdvert.html.twig', [
+            'form' => $form->createView()
+        ]);
     }
 
     #[Route('/advert/{id}', name: 'app_advert_page')]
-    public function findAdvertById(EntityManagerInterface $entityManagerInterface,Request $request,$id, AdvertRepository $advertRepository): Response
-    { 
+    public function findAdvertById(AnswerRepository $answerRepository ,QuestionRepository $questionRepository,EntityManagerInterface $entityManagerInterface,Request $request,$id, AdvertRepository $advertRepository): Response
+    {
+        /** gestion du formulaire pour ajouter une question à l'annonce */
         $question = new Question();
         $advert = $advertRepository->findOneById($id);
         $user = $advert->getUserRegister();
         $questions = $advert->getQuestions();
-        $question = new Question();
         $questionForm = $this->createForm(QuestionType::class, $question);
         $questionForm->handleRequest($request);
         if($questionForm->isSubmitted() && $questionForm->isValid()){
-            $question = $questionForm->getData();
-            $question->setAdvert($advert);
-            $question->setUserRegister($this->getUser());
-            $entityManagerInterface->persist($question);
-            $entityManagerInterface->flush();
+            //dd($questionForm->getData());
+            $question->setTitle($questionForm->getData()->getTitle())
+                        ->setUserRegister($this->getUser());
+                        
+            $advert->addQuestion($question);
+            $entityManagerInterface->persist($advert);
+            $questionRepository->save($question, true);
+
+            return $this->redirectToRoute('app_advert_page', [
+                'id' => $id
+            ]);
         }
+
+        /** gestion du formulaire pour répondre à une question */
+        $answer = new Answer();
+        $answerForm = $this->createForm(AnswerType::class, $answer);
+        $answerForm->handleRequest($request);
+        if($answerForm->isSubmitted() && $answerForm->isvalid()){
+            $questionId = $answerForm->get('questionId')->getData();
+            if($questionId){
+                $answer->setAnswer($answerForm->getData()->getAnswer())
+                        ->setQuestion($questionRepository->find($questionId))
+                        ->setUser($this->getUser());
+                $answerRepository->save($answer, true);
+                return $this->redirectToRoute('app_advert_page', [
+                    'id' => $id
+                ]);
+            }
+        }   
+
+        
+        /** vérification de la présence de l'annonce */
         if (!$advert) {
             throw $this->createNotFoundException("L'annonce que vous recherchez n'existe pas :'(");
         } else {
@@ -109,16 +132,21 @@ class AdvertController extends AbstractController
                 'advert' => $advert,
                 'questions' => $questions,
                 'questionForm' => $questionForm->createView(),
+                'answerForm' => $answerForm->createView()
                 'user' => $user
             ]);
         }
     }
+
+
     
     #[Route('/advert/{id}/delete', name: 'app_advert_delete', methods: ['GET', 'POST'])]
-    public function deleteAdvert($id, AdvertRepository $advertRepository) 
+    public function deleteAdvert($id, AdvertRepository $advertRepository, QuestionRepository $questionRepository , Advert $advert) 
     {
-        $advert = $advertRepository->findOneById($id);
-        $advertRepository->remove($advert, true);
+        if($this->getUser() == $advert->getUserRegister()){
+            $advertRepository->remove($advert, true);
+        }
+        //dd($advert->getQuestions());
         return $this->redirectToRoute('app_advert');
     }
 }
